% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical value indicating whether the contestant stays with
their original choice. If TRUE, the contestant keeps their initial pick. If
FALSE, the contestant switches to the remaining door.}

\item{opened.door}{An integer (1,2, or 3) representing the door opened by
the host.}

\item{a.pick}{An integer (1, 2, or 3) representing the door initially
selected by the contestant.}
}
\value{
A single integer (1,2, or 3) inidicating the contestant's final door
selection.
}
\description{
\code{change_door()} determines the contestant's final door selection in the
game based on whether the contestant chooses to stay with their initial
selection or switch to the remaining unopened door.
}
\details{
The function gives the contestant the option to change from their initial
selection to the other door that is still closed. The fucntion will
represent the game-playing strategy as the arguement stay=TRUE or
stay=FALSE.
}
\examples{
change_door()

}
