% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if Contestant has Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer (1,2, or 3) representing the contestant's
final door choice.}

\item{game}{A vector indicating what is behind each door (e.g., "car" or
"goat").}
}
\value{
A character string depending on the final result. "WIN" if the contestant's
final pick is the car, or "LOSE" if it is a goat.
}
\description{
\code{determine_winner()} determines whether the contestant wins or loses
based on their final door selection.
}
\details{
The function determines if the door selected by the contestant contains
a goat or the car. If the selected door contains the car, the function
returns "WIN". If it contains a goat, the function returns "LOSE".
}
\examples{
determine_winner()

}
