% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A vector indicating what is behind each door (e.g., "car" or
goat").}

\item{a.pick}{An integer (1,2, or 3) representing the door initially
selected by the contestant.}
}
\value{
A single integer (1,2, or 3) indicating the door opened by the host.
}
\description{
\code{open_goat_door()} simulates the host's opening one door that was not
selected by the contestant and does not contain the car.
}
\details{
The function shows the doors the host is allowed to open which are the
doors that were not selected by the contesant and do not contain the car.
The host randomly selects one door to open. If the contestant selects the
car on the first guess the host can open either door, but if the contestant
selects a goaat the host only has one option.
}
\examples{
open_goat_door()

}
