% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{play_game}
\alias{play_game}
\title{Simulation Set-Up}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame with two rows and two columns stating the strategy ("stay" or
"switch"), and the outcome: the result of the game ("WIN" or "LOSE").
}
\description{
\code{play_game()} simulates a single round of the Monty Hall game and
evaluates the outcomes of both possible strategies of staying with the
inital door selection or switching to the remaining unopened door.
}
\details{
The function begins by creating a new game which randomly assigns the car
and goats to the three doors. The contestant then makes their inital
selection at random. The host opens one door that the contestant did not
initially select and does not contain the car. The function computes the
contestants final door selection with them either staying with their
inital pick or switching to the remaining unopened door then evaluates
if it is a "WIN" or "LOSE".
}
\examples{
play_game()

}
