% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{play_n_games}
\alias{play_n_games}
\title{Adding the Game to a Loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer specifying the number of games to simulate.}
}
\value{
A data frame contaning the results of all simulated games, with one row per
strategy ("stay" or "switch") per game and the columns showing the outcome
("WIN" or "LOSE").
}
\description{
\code{play_n_games()} runs multiple independent simulations of the Monty Hall
game and summarizes the outcomes for the "stay" and "switch" strategies.
}
\details{
The function repeatedly plays the game (n) amount of times and stores the
outcomes from each simulation in a list. The results are combined in a
single data frame.
}
\examples{
play_n_games()
play_n_games(n = 100)

}
