% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{select_door}
\alias{select_door}
\title{Contestatnt Selects a Door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A single integer (1, 2, or 3) representing the randomly selected door.
}
\description{
\code{select_door()} has contestant makes their first selection at random of
3 available doors with equal probability.
}
\details{
The function creates a numeric vector containing the values 1,2, and 3 to
represent the available doors which it randomly selects exactly one door
with equal probability of being chosen.
}
\examples{
select_door()

}
